🏗️ Architect: Modularized weather plotting system
Modularized the weather plotting system by splitting the monolithic
`apts/plotting/weather/plots.py` into a modular package structure.

- Extracted individual parameter plotting functions to `params.py`.
- Refactored individual parameter plots to use a data-driven helper
  `_generate_param_plot`, significantly reducing code duplication.
- Extracted summary-level plotting functions to `summary.py`.
- Consolidated shared helper logic, including `mark_observation` and
  `mark_good_conditions` re-exports, into `utils.py`.
- Updated `plots.py` to act as a coordinator and re-export the public API
  for backward compatibility.
- Updated `test_individual_weather_plots.py` to match the new module
  hierarchy for consistent mocking.

This refactor improves maintainability by adhering to the Single
Responsibility Principle and reducing structural complexity.
